Add TaxableAmount to LineVat and update creation logic
A new `TaxableAmount` property was added to the `LineVat` class
in the `SaphetyToDDL.Lib.Models.Saphety` namespace. It is of
type `decimal` and annotated with `[XmlElement("taxableAmount")]`
for XML serialization/deserialization.

The logic for creating a `LineVat` object in `SaphetyToDDL.cs`
was updated. The `TaxTotalValue` property is now explicitly set
to `line.TotalTaxAmount` or `0` if it doesn't exist. The new
`TaxableAmount` property is set to `line.TotalNetAmount`.